[iris] dedupe profile dashboard panels and move IrisProfile to runtime/profile
Extract a shared <ProfileHistory source="..." /> component used by
StatusTab, TaskDetail, and WorkerDetail in place of three near-identical
~100-line blocks. Move IrisProfile + the Profile{Type,Format,Trigger}
enums + PROFILE_NAMESPACE from worker/stats.py into runtime/profile.py
so build_profile_row no longer crosses the runtime → worker layer
inversion.
